# Http2 client configuration to improve performance