--- a/actix-http/src/client/config.rs
+++ b/actix-http/src/client/config.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 // These values are taken from hyper/src/proto/h2/client.rs
 const DEFAULT_H2_CONN_WINDOW: u32 = 1024 * 1024 * 2; // 2mb
-const DEFAULT_H2_STREAM_WINDOW: u32 = 1024 * 1024 * 1; // 1mb
+const DEFAULT_H2_STREAM_WINDOW: u32 = 1024 * 1024; // 1mb
 
 /// Connector configuration
 #[derive(Clone)]