--- a/actix-web-codegen/Cargo.toml
+++ b/actix-web-codegen/Cargo.toml
@@ -12,7 +12,7 @@ workspace = ".."
 proc-macro = true
 
 [dependencies]
-quote = "^1"
+quote = "=1.0.2"
 syn = { version = "^1", features = ["full", "parsing"] }
 proc-macro2 = "^1"