This PR should be fixing https://github.com/actix/actix-web/issues/1393. 

Since implementation is using hardcoded const need your opinion (pls see question below). Implementation is pretty much based on [hyper's resolving similar problem](https://github.com/hyperium/hyper/blob/master/src/proto/h2/client.rs#L24-L28), though it sets initial buffer for quite large 5MB which gets some pressure off a task scheduler but requires more memory.

UPD: this PR also adding possibility to configure awc and Connector:
- initial_window_size
- initial_connection_window_size
- min_http_version

I have pointed awc_https example to this branch to show that it really helps:
https://github.com/dunnock/examples/tree/h2-speed-1393
Also it can be seen from benchmarks 
https://github.com/dunnock/awc_test/tree/benches

Speed with enabled alpn protocol:
```
	let mut builder = SslConnector::builder(SslMethod::tls()).unwrap();
	builder.set_alpn_protos(b"\x02h2\x08http/1.1").unwrap(); 
```
```
Benchmarking awc_test_openssl: 
Throughput 5994288B/s
Iters 10 total 101747060 elapsed 16974ms
awc_test_openssl        time:   [1.5834 s 1.6329 s 1.6794 s]                              
```
```
Benchmarking awc_test_default
Throughput 6327160B/s
Iters 10 total 101747060 elapsed 16081ms
awc_test_default        time:   [1.5967 s 1.7657 s 2.0360 s]                              
```
compared to reqwest
```
Benchmarking reqwest:
Throughput 6333067B/s
Iters 10 total 101747060 elapsed 16066ms
reqwest                 time:   [1.6207 s 1.7328 s 1.8409 s]                     
```