@Dowwie 2020-03-03
Great job!  This seems to have fixed the 10MB download issue.