@dunnock 2020-03-04
> Looks good! Could you also update changelog?
Done