@dunnock 2020-03-04
On my question I also looked into implementation of adaptive window and it seems an overkill: for every poll it uses Mutex to store calculation. Maybe that's the reason why it's disabled by default in hyper client.

I am thinking of adding 3 parameters to the ClientBuilder as part of another PR:
- h2_initial_window_size
- h2_initial_connection_window_size
- h2_enabled -- one off requests would benefit of using HTTP1.1
what do you think?