@robjtede 2020-03-04
More options are always beneficial.

With HTTP3 on the horizon, I'd argue to drop the `h2_` prefixes and use a more forwards compatible option for `http1_only`; something like `http_version` to declare the max "supported" protocol.