@Dowwie 2020-03-04
useful feedback from gitter channel:

```
Grant Jennings @gajbooks 20:54:
Actually, it is not an issue, but a misconfiguration in the usage. By default, the window size is set to 65k both on the connection level and on the stream level in h2. This results in a packet being sent every time 65k of data is received, greatly increasing latency.

Grant Jennings @gajbooks 21:03
For sub-65k payloads this would not be an issue, but it heavily caps throughput for larger files. H2 does not have built-in flow control adjustment.

Grant Jennings @gajbooks 21:09
It is also not specified in the spec, and likely will never be implemented in h2 as a result.
To pick a random number, 1MB seems to give very good performance, even over .5MB. However, anything over 1MB doesn't improve performance. (Not any sort of sane statistical test, just anecdotal evidence.)
```