@dunnock 2020-03-04
Based on this feedback I will try to test if performance same with 1MB window.
That would be easy to test when I add configurable parameters, so probably will extend this PR with configuration, should be ready soon...