@dunnock 2020-03-04
I have tested with variable buffer size: 500kb, 1MB and 2MB. 2MB gives little benefit over 1MB (10% for big downloads), 500kb 25% slower than 1MB. Based on this I have set defaults to 1MB stream buffer and 2MB connection buffer.
```
awc_test_default(1MB)        time:   [1.9620 s 2.0073 s 2.0645 s]
awc_test_window_2mb     time:   [1.7536 s 1.7912 s 1.8353 s] 
awc_test_window_500kb   time:   [2.3543 s 2.5123 s 2.6838 s] 
reqwest                 time:   [1.8177 s 1.8597 s 1.8981 s]   
```

This is now ready for final review with parameters configuration to awc::ClientBuilder and actix-http::client::Connector.