@dunnock 2020-03-05
This poining to the common build problem https://users.rust-lang.org/t/failure-derive-compilation-error/39062