@JohnTitor 2020-03-05
Let's pin quote version in this PR, could you?